pyproject.toml
[project]
name = "ocr-reader"
version = "0.0.1"
description = "一个用于图片处理、OCR识别和翻译的工具"
readme = "README.md"
requires-python = ">=3.9, <3.10"
dependencies = [
    "opencv-python==4.9.0.80",
    "numpy<2.0",
    "pillow",
    "pyqt5==5.15.10",
    "pyqt5-qt5; sys_platform != 'win32'",
    "pytesseract",
    "sentencepiece",
    "torch==2.2.2",
    "transformers",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
